arch: arm: Move header files to common location
These files will be used for Cortex-R support as well.
