+++ CODEOWNERS
@@ -223,7 +223,7 @@
 /include/arch/arc/arch.h                  @andrewboie
 /include/arch/arc/v2/irq.h                @andrewboie
 /include/arch/arm/                        @MaureenHelm @galak @ioannisg
-/include/arch/arm/cortex_m/irq.h          @andrewboie
+/include/arch/arm/irq.h                   @andrewboie
 /include/arch/nios2/                      @andrewboie
 /include/arch/nios2/arch.h                @andrewboie
 /include/arch/posix/                      @aescolar
