+++ include/arch/arm/misc.h
